Name changes

+++ src/Ramp.cpp
@@ -1,12 +1,12 @@
 /* RAMP.CPP
     This module defines the rampClass, which implements a single trapezoidal ramp function where the ramp-up and -down durations are equal. 
-    The class variable val is used to access the current ramp function value, which is 0 at the ramp start, 1 during the hold phase, and 0
+    The class variable (float) val is used to access the current ramp function value, which is 0 at the ramp start, 1 during the hold phase, and 0
     again at the end of the ramp-down. The ramp value is typically used to scale the output of some other effect, to provide a smooth 
     transition into and out of the effect.
     The ramp-up/down durations (for all subsequent ramps) can be set using setRamp() or using one of the overloaded start() functions. 
 */
 #include <Arduino.h>
-#include "MotionUtils.h"
+#include "EffectUtils.h"
 #include "Ramp.h"
 
 
@@ -44,7 +44,7 @@ void rampClass::start(float duration) {
   float rampTimeNow;    // potentially-constrained ramp time for this start instance
 
   if (duration == 0) {  // if ramp has infinite duration
-    rampSteps = ComputeSteps(rampTime);  // number of steps in up or down ramp
+    rampSteps = ComputeSteps(rampTime);  // number of steps in ramp-up
     holdSteps = 0;  // hold phase has infinite duration after ramp up
   }
   else {  // finite duration ramp
